+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Bug when home group does not exist
+- Background color for time selection in the searching reporting and tooltip text color in dark theme
 
 ## [1.115.0]